Made delete thread even less dependent on the current state / state changes.
Nodes were somehow getting lost, with osci-manage --list showing the node
as not in use, but it was still held by nodepool.

With this change we now use nodepool to list the nodes that we can delete.
Also changed from lists to sets as nodes cannot appear more than once
and set arithmatic is nice.